Add files via upload
You will need to have MYSQL workbench to run this project easily, open mysql workbench create a schema named evenbatch and  then just in the application properties "remove # for the first time" then RUN IT as spring boot App, this is done so that your database gets automatically created. Once this is done Again add # as earlier in the application properties ,so that the database doesn't get created every time you run the application. In this,
You will be able to perform CRUD operations on a Employee form by running it on localhost:8080.